Added support for using image files as input
The application is now capable of processing an input image file of any
type supported by OpenCV. The file is specified in the command line in the
same way of video files: the application looks at the file extension to
decide if the specified parameter is a video filename or an image
filename.

Processing of images is done in just one step, as there is only one frame
to process.